Added stalebot. (#16)
Set to trigger a message on stale issues after 30 days, if there is no response 14 days after that, it is automatically closed with the wontfix label. The exceptions to this are issues marked with security and pinned labels. The pinned label indicates issues or bugs that we don’t want to go stale (i.e., ones that have been added as part of milestones/roadmap).