Disable test_020.php for uopz >= 7.1
The uopz integration test case test_020.php relies on the behavior of
the statement eval('exit(10)') in order to create an exit code so it
can be examined with uopz_get_exit_status().  For PHP 8 and uopz 7.1
this does not appear to work any longer so disabling the test for now.
Needs to be examined and fixed if possible.