fix(daemon): Fixes to handling communications with collector
This work is based on the current PR for adding P17 support for
handling of response codes from the collector.  It contains some
fixes for handling the body of these responses properly (only
trying to unmarshal the 'return_value' JSON entity if the
resposne code is 200) as well making sure errors during the
communication with the collector are properly propagated up
in the ConnectAttempt struct.