feat(agent): instrument Monolog logging (#489)
* Instrument Monolog library to forward logs to New Relic

`Supportability/Logging/PHP/Monolog/enabled`
is created when Monolog logging library is detected.

Additionally `Supportability/library/%s/detected` metric is created for
detected logging frameworks.

Instrument `Monolog\Logger::addRecord` to capture $level, $message, 
$context and $datetime (if available) to create log event for New Relic.
The message of log event is constructed by concatenation of $message 
and  $context, if it is not empty array. If $context is an empty array 
it will not be appended to the final message.

* Detect conflict with `NewRelic\Monolog\Enricher`

Log a warning when `NewRelic\Monolog\Enricher\Handler` is detected in
`Monolog\Logger`'s handler stack.

* Create logging metrics

Create/update Logging/lines and Logging/lines/{severity} metrics for
every log record added. This is done only if enabled by configuration.

Co-authored-by: Michael Fulbright <89205663+mfulb@users.noreply.github.com>