feat(agent): Convert transactions to use nr_log_events_t
In this implementation using priority sampling the previous use of
nr_analytics_events_t is removed and replaced with the new
heap based implementation.  This requires changes such as
assigning a priority to each log event.  The choice here is to use
the priority flag from the corresponding segment.  This mirrors
how spans are given priorities.