feat(agent/daemon): Implement log event limits configuration (#497)
* feat(agent/daemon): Update flatbuffers for flatc v2.0.6

This change is mostly just updating the autogenerated files from
flatc to accomodate v2.0.6 of this compiler.

I also added a entry in the App table for the log event max limit.

* feat(agent/daemon): Add max log events limit to App table

* feat(agent/daemon): Add log_events_max_samples_stored to flatbuffer

Added log_events_max_samples_stored to the "App" table in the
flatbuffer defintion.  This allows the agent to send its idea
of the max log event limit to the daemon, which will then
send this value in the connect request to the collector.

* feat(agent/daemon): Share log event limits between agent/daemon

Leverages modifications to protocol.fbs to send the log event
limit from the agent to the daemon.  The daemon uses this
value when creating the connect request for the collector.

* feat(agent): Initializes log event limit at transaction creation

* test(agent): Add tests for log event limits communcations

* feat(daemon): Implement log event limits configuration

Implements changes to support accepting a log event limit
from the agent (from INI config) and passing it in a
collector connect request.  The collector response is
parsed and the collector suggested value for the log limit
it retrieved and compared to the agent requested log limit.
The smaller of these values is then used for setting up
new harvest cycles.

* feat(daemon): Add handling of log limits and tests

* fix(daemon): Rework handling invalid input value

* style(daemon): Remove debug comment

* style(daemon): Remove duplication code from rebase

* fix(daemon): Remove duplicate assignment

* test(daemon): Add agent limit for log events so test will run

* test(daemon): Remove duplicate assignment

* style(daemon): Properly format Go source

* style(agent/daemon): Cleaned up troubleshooting debug messages

* fix(daemon): Rework log event limit w/o a default

* fix(daemon): Refactor dupe code into help function

* chore(daemon): Fix spelling error.

Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>

* fix(test): Use better function name and remove unneeded code

* fix(daemon): Add debug msg if agent limit for logs selected

* fix(agent): Enforce smallest log event limit in agent

Need to enforce the smallest of the collector harvest limit and the agent
config INI value here.  This is in part because the enforcement in the daemon
will not be exercised in integration tests due to how the integration_runner
doesn't follow the same code path when processing a connect response
as in the normal operation of the daemon.

* fix(agent): Format uint64_t type prperly

Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>