test(logging): Fix daemon test for log harvest
Currently the daemon removes the first and last character of a
json log event representation from the agent because the agent
wraps each one with "[" "]" to make it a valid json string.  The
collector does not want these characters present when created
the final payload of all log events.  So the sample log event for
this test case has had the "[" "]" added.
Also the expected json from a log event collector payload was
updated to the proper format following the spec.