fix(agent): Remove unneeded log event functionality
This commit removes the function nr_log_event_to_json_buffer_ex()
as it is not needed.  Originally the idea was to have a function to
produce a more compact JSON representation but after review
the compact representation is adequate for all use cases.  This
makes the function nr_log_event_to_json_buffer_ex() unnecessary.

Code using this function as well as affected test cases have been
updated.