test(agent): Improve log events test
Due to the uncertainty of the order events pulled from a
minmax heap via an iterator some test cases needed to
be reworked.  Instead of testing the internal construction
of the minmax heap the test case verifies the correct
events have been inserted.
The operation of the minmax heap is currently better
exercised in the test_events_sample() test which inserts
more events in the heap than it can hold and verifies
the highest priority events are retained.