Fix `NRPRG(php_cur_stack_depth)` updates in OAPI
NRPRG(php_cur_stack_depth) is used to keep track of PHPs stack depth.
It is therefore unconditionally incremented when the agent enters
nr_php_execute and unconditionally decremented before function returns.
This logic needs to be preserved when PHP's observer API is used for
instrumentation. NRPRG(php_cur_stack_depth) must be decremented in
func_end callback regardless if agent is or is not recording transaction
or it will not correctly reflect stack depth.