test: set LogEventConfig in integration_runner (#571)
Hardcode 10000 as the integration_runner's daemon default log event
harvest limit that is to be used when connecting to collector.

integration_runner does not use agent settings from integration tests
for values in pre-connect/connect payload. It is using pre-cooked app
info structure, and any uninitialized fields in that structure, like
AgentEventLimits.LogEventConfig.Limit, which is sent in connect payload
as the value for event_harvest_config.harvest_limits.log_event_data,
gets the default value of 0. Collector accepts and honors this value,
which results in integration_runner's daemon and agent to be configured
not to collect any log events. This breaks logging integration tests.