Fix creating some logging metricss
The logging metrics "Supportability/Logging/Forwarding/PHP/<state>" and
"Supportability/Logging/Metrics/PHP/<state>"} are supposed to be created
once per request.  The previous mechanism did not properly create these
metrics if a transaction was started and restarted during a request, or
if newrelic_set_appname() was called.

This fix uses a flag within the transaction structure to track if the
metrics have been created for the transaction which fixes the previous
issue.