refactor(agent): user instrumentation lookup
Speed up user instrumentation lookup with the use of hashmap that stores
pointers to wraprecs that have instrumentation enabled. The hashmap
key generation method is based on nr_php_function_debug_name:

- for user function: combine scope (if any) with function name
- for closure: cobine filename with line number

This guarantees uniqueness in most cases. The key will only be the same
for closures declared on the same line in the same file.