feat(agent): Add code level metrics functionality. (#585)
* feat(agent): Add code level metrics functionality. 
1) Add INI value to disable/enable code level metrics. 
2) Create new function `nr_php_txn_add_code_level_metrics` that uses
nr_php_zend_execute_data_* family of functions to extract CLM and add it
as an `agent attribute`.
3) Added new NR_TXN_ATTRIBUTES for `code.namespace`, `code.lineno`,
`code.filepath`, and `code.function`.
4) Added call to php_execute_enabled to call the new CLM function 
5) Added new integration tests to exercise the new functionality. 
6) updated `nr_php_execute_metadata_t` to hold additional information
and moved the definition of the struct to php_execute.h.
7) `nr_php_execute_enabled` uses the metadata at the beginning so now we
can simply release at the end.
8) In the case of a super short segment that would have been ignored, we
do NOT add CLM (otherwise it wouldn't be ignored). We wait until after
we decide to ignore or not to add the CLM data.
9) check clm string lengths so don't add attributes if the string will
be truncated

Note:  CLM functionality is only compatible with PHP 7+.

10) tailored CLM to current metadata utilizing zrefs instead of strdups;
otherwise, it was adding too much overhead.