fix: replace hashmap implementation
Replace a generic hashmap (using zend_function address as a key) with
specialized hashmap that uses zend_function's metadata (filename,
lineno, scope name and function) a key and wraprec as a value because in
PHP 8.1 zend_function address can change within a single request between
when it's added to the hashmap and when it's being looked up in the
hashmap.

In order to improve performance of the specialized hashmap key matching,
zend_function's metadata is stored in a key as a refcounted zend_string
rather than raw char*. This allows key matching to take advantage of
following zend_string attributes: length (len) and hash (h) - both
values are pre-calculated by Zend engine and available for use by the
agent.  Because the agent instruments only a small fraction of all user
functions in a PHP application, there will be more missed hashmap
lookups than hits. Therefore delaying string value comparison until
absolutely necessary (len and h must match match) is a significant
performance gain.