test: add tests to cover two funcs with same name
1. wrap two functions in the same file with same name but different
class and verify the set/get
2. wrap two functions in different file with the same name and different
class and verify set/get

Following tests were considered:
3. wrap two functions in different file with the same name and no class
and verify set/get
4. wrap two functions in different file with the same name and same
class and verify set/get
This is not allowed in PHP within a single transaction. I.e. if PHP was
to execute this type of code, it will complain that the function is
already defined in previous file. If this however happens (I’m not
really sure how :idk:), the way lookup hashmap is implemented it will
only store the latest wraprec.