fix(agent): Add more stringent opcache checking. (#595)
More stringent checks are required for PHP 8.1+ prior to querying
opcache to prevent warning messages from being thrown. Additionally, the
severity level of the messages should be downgraded to `debug`.

1) If `opcache.enable`/`opcache.enable_cli` is true and `opcache.preload` is set by user then
call the attempt to instrument the preloaded functions.
2) According to
https://www.php.net/manual/en/function.opcache-get-status.php will
return either an array or false. Hence
https://github.com/newrelic/newrelic-php-agent/blob/main/agent/php_execute.c#L1690
will be either an array or false
both of which could be valid depending on the state of PHP.
Therefore the warning message:
`User instrumentation from opcache: opcache status information is not an
array` is downgraded to debug.
3) clang format picked up a few changes.

Passes unit and integration tests.
Tested locally for functionality and for error msg using slim soak
tests.