refactor(agent): user instrumentation lookup (#588)
A more performant data structure is used to lookup user instrumentation:
a hashmap with zend function's metadata (lineno, functioname, scope
and filename) as a hash key, and a pointer to a wraprec as a value.

In order to improve performance of the specialized hashmap key matching,
zend_function's metadata is stored in a key as a refcounted zend_string
rather than raw char*. This allows key matching to take advantage of
the following zend_string attributes: length (len) and hash (h) - values
for
both are pre-calculated by Zend engine and available for use by the
agent.  Because the agent instruments only a small fraction of all user
functions in a PHP application, there will be more missed hashmap
lookups than hits. Therefore delaying string value comparison until
absolutely necessary (len and h must match first) is a significant
performance gain.

To save time on lookups, a single hashmap will be used to store pointers
to both types of wraprecs: transient and reusable. For efficiency, the
wraprecs that are re-usable between requests (e.g. for well known
library or framework user functions), are additionally stored in a
linked list.  These wraprecs are created once per interesting function
detection, and destroyed at module shutdown. Transient wraprecs, that
are not re-usable between requests, are not stored in linked list.
Transient wraprecs are created on the fly and destroyed at request
shutdown.  However user function instrumentation installation (aka
wrapping) is done the same way for both types of wraprecs and happens
once per each request, i.e. for each request the hashmap is created
anew, when user function is instrumented, its wraprec is added to the
hashmap, and at the end of the request the hashmap is destroyed together
with transient wraprecs. Re-usable wraprecs are not destroyed - they
remain in linekd list and are re-set (marked as not wrapping a zend
function). This happens because with new request/transaction php is
loading all new user code.

Co-authored-by: Michael Fulbright <89205663+mfulb@users.noreply.github.com>