feat(agent): Remove php_vm.c overwriting opcodes when using OAPI (#559)
Occasionally, we want to instrument functions called with
call_user_func_array, when we otherwise wouldn't instrument that
function. To do this, we instrument the internal function
call_user_func_array and check for contexts in which we want to begin
instrumentation. However, call_user_func_array can be inlined by the
zend compiler. Previously, we were detecting inlined calls by
overwriting the DO_FCALL opcode. With OAPI, we no longer want to touch
the opcodes and are instead checking for the inlined calls during
observer_do_fcall.