fix(agent): Ensure txn exists before we try to reference it. (#600)
Consistent with other functions in php_execute, check if txn exists
before trying to use it.

The issue occurred during the first request, and because it is the
first, the appname is unknown and we don't create a txn. However, an
exception happened in that first request, our exception_hook handler got
triggered, and if that is triggered, we always try to close off dangling
segments. In this case, because we didn't check for txn==null, it
segfaulted as we tried to access txn elements.
Pr fixes that in two ways, 1) check for txn==null before closing off
segments. 2) don't record uncaught exception info generated from our
exception_hook handler if a txn is null.
```
2023-01-26 21:03:16.816 +0000 (29813 29813) verbosedebug: RINIT processing started
2023-01-26 21:03:16.816 +0000 (29813 29813) debug: added app='PHP Test Apps Laravel ads v2' license='07...4a'
2023-01-26 21:03:16.816 +0000 (29813 29813) verbosedebug: querying app='PHP Test Apps Laravel ads v2' from parent=4
2023-01-26 21:03:16.816 +0000 (29813 29813) verbosedebug: sending appinfo message, len=6492
2023-01-26 21:03:16.817 +0000 (29813 29813) debug: APPINFO reply unknown app='PHP Test Apps Laravel ads v2'
2023-01-26 21:03:16.817 +0000 (29813 29813) debug: unable to begin transaction: app 'PHP Test Apps Laravel ads v2' is unknown
```
note the last line: debug: 
unable to begin transaction: app 'PHP Test Apps Laravel ads v2' is
unknown