Refactor release-installer target
Separate release-installer-script and release-installer-iutil targets
from release-installer target. This is needed because newrelic-install
is not a binary artifact, which does not depend on platform arch, and
newrelic-iutil is a binary artrifact, which depens on platform arch.
Separating these targets out simplifies building agent release when
using containers as build platform.