link to static libpcre for release build
Use PCRE_STATIC environment variable to enforce linking to static libpcre
to ensure the agent is able to run on all Linux distros. This is needed
because different Linux distributions use different names for libpcre
shared object and therefore its impossible to link to a shared object
that works universally on all Linux distributions.