fix(agent): Combine the nr_php_call 8.2 and 8+ blocks. `zend_call_method_if_exists` actually had the affect of suppressing a valid warning message.
When we ran it, because we forced Wordpress as a framework, we made a call in fw_wordpress.c which makes a call to get_theme_roots.
However, in the test case
get_theme_roots (which normally exists in Wordpress) does not exist in our "mocked" Wordpress so we received a perfectly valid error msg.

Fatal error: Uncaught Error: Invalid callback get_theme_roots, function "get_theme_roots" not found or invalid function name in /usr/local/src/newrelic-php-agent/tests/integration/frameworks/wordpress/test_wordpress_apply_filters.php8.php:66
Stack trace:
  thrown in /usr/local/src/newrelic-php-agent/tests/integration/frameworks/wordpress/test_wordpress_apply_filters.php8.php on line 66

If we use nr_php_call we don't want it suppressing exceptions that could be useful to us or the user.

The error message is good, the problem is that after this, we weren't detecting that something invalid happened and were copying/returning retval even though we shouldn't.
The fix is to
a) check retval for undef and if so, return NULL.
b) add get_theme_roots to the test case.

2) initialize retval to prevent segfaults.
zval retval = {0};