feat(integration_runner): Add PHPMODULES section
This commit adds a PHPMODULES section which allows the following:

/*PHPMODULES
zend_extension=opcache.so
extension=mysqli.so
*/

The modules are only loaded when the test case is run and not when the
SKIPIF code snippet is run.  This was necessary to avoid problems seen
running test scripts written against newer PHP versions on older PHP
versions which resulted in syntax errors on newer language features.