feat(integration_runner): Add 'warn' test result
This commit adds a 'warn' test result which is intended to be used to
represent tests which could not be run due to something not being
correct in the environment.  The intention is 'skip' is for tests which
cannot be run because of the PHP version or some other expected
condition which would make the test not run in all environments.  The
use of 'warn' is, for example, if a PHP extension was not present which
is required and if missing means the test environment is mis-configured.
The new integration_runner boolean option 'warnisfail' will cause any
warnings to be treated as a failure and the integration_runner will exit
with a non-zero exit code to signify the failure.