feat(daemon): Add loglevel for daemon health (#641)
This PR adds a new log level for the daemon called "healthcheck". It is
between "debug" and "info" in verbosity.
If enabled this log level will currently output a daemon log message of
type "HealthCheck" which will output the current number of applications
connected to the agent as well as the maximum number of allowed
applications.
The format of the line is:
```log.Healthf("current number of apps is %d of a max of %d", numapps, limits.AppLimit)```
where ```numapps``` is the current number of applications and ```limits.AppLimit``` is the compile-time limit for the number of applications a daemon instance will allow.