feat(integration_runner): Add 'warn' test result (#642)
This commit adds a 'warn' test result which is intended to be used to
represent tests which could not be run due to something not being
correct in the environment.

The intention is 'skip' is for tests which cannot be run because of the
PHP version or some other expected condition which would make the test
not run in all environments. However, currently 'skip' is used to also
represent when a test which is expected to run could not run due to a
misconfiguration of the environment. This can cover up potential test
run deficiencies.

The use of 'warn' means the test environment is mis-configured. 

A new integration_runner boolean option 'warnisfail' will cause any
warnings to be treated as a failure and the integration_runner will exit
with a non-zero exit code to signify the failure.

The 'warn' result can be returned by doing something like:

 ```die("warn: could not load important thing so test not run")```.