fix(agent): disable CAT when DT enabled (#658)
### Internally disable the global for CAT when both CAT and DT are
enabled

This PR will throw a warning message in the agent log about the CAT + DT
conflict, suggest to the end user that they modify the setting, and will
modify the `cross_process_enabled` INI value for the duration of the
transaction to safeguard against any potential holes in the code that
could allow both CAT and DT functionality present in the same
transaction.

This PR does not modify the INI file, and checking INI settings (with
`phpinfo()` or `php -i`) will not show this change reflected. As far as
I'm aware, that's only possible by modifying the INI file itself, which
is not something we should do.

A test has been added to verify this scenario.