separate agent service into php and devenv
php service is built using nr-php-agent-builder image and is to be used for
running integration tests as part of pull request test. devenv service is
built using Dockerfile and is to be used for local development. devenv service
is built using Dockerfile to make dockerized development environment available
to external contributors until nr-php-agent-builder image is made public.