ci: add on-demand run-integration-tests workflow (#680)
An on-demand workflow that will run integration tests for PHPs 8.0, 8.1
and 8.2 on gnu and musl platforms on x86_64 and aarch64 architectures.
This workflow runs integration tests only, because unit tests, as well as
custom events limits tests, and LASP tests, are already part of the
test-pull-request workflow.

It is expected that running tests on aarch64 architecture will cause
test failures, because tests run slower on emulated platform and more
spans are generated. aarch64 environment needs some adjustments.