ci: fix on-demand run-integration-tests workflow (#688)
Fix `experimental` flag in strategy matrix of on-demand
run-integration-tests workflow. It is expected that running tests on
aarch64 architecture will cause test failures because tests run slower
on emulated platform and more spans are generated and therefore this
arch is marked as `experimental` and failed tests on that arch will not
mark the whole workflow as failed.