fix: add guards to oapi cufa opcode check (#708)
When instrumenting wordpress/drupal hooks, we want to instrument the
hook function calls, which are dispatched by php's
call_user_function_array. This php function (cufa for short) is
flattened/inlined, so to determine if we are in a function called by a
cufa call, we need to check the opcodes of the caller to the current
function. If the caller is a php internal function, checking these
opcodes was not safe and lead to invalid reads. As such, before reading
the opcodes, we first check if the caller is a user function. We only
want to instrument cufa calls from user functions anyway.

---------

Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>