add tests for version 7 of guzzle
Version 6 of Guzzle supports PHPs up to but not including 8.0. In order to test
agent's instrumentation of guzzle on PHPs 8.0+, version 7 of Guzzle needs to
be used. There are no fundamental changes in version 7 of Guzzle therefore
agent's instrumentation of version 6 of guzzle6 works for version 7 too. However,
this results in guzzle7 being detected as guzzle6 and that's why metrics in
integration tests expectation include `Supportability/library/Guzzle 6/detected`
as well as `Supportability/library/Guzzle 4-5/detected`.

guzzle.tar.bz2 bundle has been updated to include version 7 of Guzzle.