tests(agent): enhance guzzle tests (#713)
* cleanup skipif.inc for guzzle5: 
skipif.inc for guzzle5 ensures tests run on correct version of PHP
and skip test if PHP version does not fall into >=5.4,<7.4 interval.
Additionally, it warns (and skips) if guzzle5 installation is not found.

* add skipif.inc for guzzle6
skipif.inc for guzzle6 ensures tests run on correct version of PHP
and skip test if PHP version does not fall into >=5.5,<8.0 interval.
Additionally, it warns (and skips) if guzzle6 installation is not found.

* add tests for version 7 of guzzle
Version 6 of Guzzle supports PHPs up to but not including 8.0. In 
order to test agent's instrumentation of guzzle on PHPs 8.0+, 
version 7 of Guzzle needs to be used. There are no fundamental
changes in version 7 of Guzzle therefore agent's instrumentation
of version 6 of guzzle works for version 7 too. However, this results
in guzzle7 being detected as guzzle6 and that's why metrics in
integration tests expectation include `Supportability/library/Guzzle
6/detected` as well as `Supportability/library/Guzzle 4-5/detected`.

* composer update guzzle 6 & 7 in guzzle.tar.bz2
  - Upgrading guzzlehttp/guzzle (6.3.3 => 6.5.8) (with dependencies)
  - Upgrading guzzlehttp/guzzle (7.4.5 => 7.7.0) (with dependencies)