chore: EOL support for build and dist for PHP 5.x (#720)
Support for PHP 5.x ended with v10.11.0.3 release. Therefore installer
script is updated to not install the agent on the system with PHP 5.x
now consider unsupported. Additionally build system is no longer capable
of generating artifacts for PHP 5.x. Moreover, to optimize build time,
the agent will no longer be built for PHP 5.x in the following
workflows:
- `code-coverage-baseline`
- `test-pull-request`