improve skipped tests handling
PHPUnit passes skipped tests to `addError` as well as `endTest` regardless of
the version. The difference is that in PHPUnit 5.x it is not possible to obtain
test outcome for skipped test - `$test.getStatus()` returns 'null'. Because of
that, `endTest` instrumentation bails out and does not generate test event and
therefore `addError` needs to be instrumented. It is possible to obtain test
outcome for skipped tests in PHPUnit 9.x. However, the `$time` for skipped test
is not received in `endTest` as a valid double and therefore it needs to be
'fixed'. Moreover testcase.getStatusMessage returns 'null' for tests skipped due
to dependency failure and therefore it is not possible to expect it in the test
event.